+++ webextension/js/classes/chrome-native.js
@@ -7,12 +7,12 @@ import {
 } from "./chrome-native-settings.js";
 import {sendNotification} from "./chrome-notification.js";
 import {getCurrentTab} from "../utils/getCurrentTab.js";
-import {tabPageServerIpStorage} from "../variousFeatures/tabPageServerIp.js";
 import ipRegex from "../../lib/ip-regex.js";
 import {io} from "../../lib/socket.io.esm.min.js";
 import {isServiceWorker} from "../utils/browserDetect.js";
-import {updateStyles} from "../variousFeatures/contentStyles.js";
+import {contentStyles, updateStyles} from "../variousFeatures/contentStyles.js";
 import {contentScripts} from "../variousFeatures/contentScripts.js";
+import {_tabStylesStoreKey} from "../constants.js";
 
 
 /**
@@ -123,6 +123,10 @@ socket.on('ws open', function (err) {
 		.catch(console.error);
 });
 
+socket.on('doRestart', function () {
+	chrome.runtime.restart();
+});
+
 socket.on('disconnect', function (reason, description) {
 	console.log('[NativeMessaging]', 'ws close', reason, description);
 
@@ -228,7 +232,7 @@ socket.on('onSettingUpdate', function (preference) {
 
 
 chrome.storage.onChanged.addListener(async (changes, area) => {
-	if (area === 'session' && tabPageServerIpStorage in changes) {
+	if (area === 'session' && _tabStylesStoreKey in changes) {
 		sendSocketData()
 			.catch(console.error)
 		;
@@ -332,14 +336,25 @@ async function sendSocketData() {
 		.catch(console.error)
 	;
 	if (activeTab) {
-		const raw = (await chrome.storage.session.get([tabPageServerIpStorage])),
-			data = Object.assign({}, raw[tabPageServerIpStorage])
-		;
+		try {
+			/**
+			 *
+			 * @type {ContentStyles}
+			 * @private
+			 */
+			const _contentStyles = await contentStyles;
+			const _tabData = _contentStyles.tabData[activeTab.id.toString(36)];
+			if (_tabData && _tabData.customData && _tabData.customData.requestDetails) {
+				tabData = _tabData.customData.requestDetails;
+				tabData.pageRating = _tabData.customData.metaRating;
+			}
+		} catch (e) {
+			console.error(e);
+		}
 
 		/**
 		 * @type {undefined|TabPageServerIdData}
 		 */
-		const _tabData = data[`${activeTab.id}`];
 		let url, domain;
 		try {
 			url = new URL(activeTab.url);
@@ -386,14 +401,12 @@ async function sendSocketData() {
 		tabData = {
 			name: activeTab.title,
 			faviconUrl: favicon,
-			error: _tabData?.error ?? undefined,
-			statusCode: _tabData?.statusCode,
+			statusCode: tabData?.status,
 			url: activeTab.url,
 			domain,
-			ip: _tabData?.ip,
+			ip: tabData?.ip,
 			ipMore,
-			openGraph: _tabData?.tabOpenGraphData ?? undefined,
-			pageRating: _tabData?.pageRating ?? undefined,
+			pageRating: tabData?.pageRating ?? undefined,
 		}
 	}
 
@@ -644,6 +657,22 @@ export async function writeClipboard(data) {
 	return await socket.timeout(timeout).emitWithAck('writeClipboard', data);
 }
 
+/**
+ *
+ * @param {string} templateName
+ * @param {object} context
+ * @returns {Promise<string>}
+ */
+export async function nunjuckRender({templateName, context}) {
+	const data = await socket.timeout(timeout).emitWithAck('nunjuckRender', templateName, context);
+	if (data.error) throw new Error(data.error);
+	if (!('result' in data)) {
+		throw new Error(JSON.stringify(data, null, "\t"));
+	}
+	return data.result;
+}
+
+
 
 
 
@@ -688,6 +717,22 @@ chrome.runtime.onMessage.addListener(function (message, sender, sendResponse) {
 			.then(() => {
 				sendResponse({
 					isError: false
+				});
+			})
+			.catch(err => {
+				console.error(err);
+				sendResponse({
+					isError: true
+				});
+			})
+		;
+		return true;
+	} else if (message.id === 'nunjuckRender') {
+		nunjuckRender(...message.data)
+			.then((data) => {
+				sendResponse({
+					isError: false,
+					response: data,
 				});
 			})
 			.catch(err => {