+++ webextension/js/init-templates.js
@@ -1,9 +1,20 @@
+/**
+ *
+ * @param {string} templateName
+ * @param {object} context
+ */
+export function nunjuckRender(templateName, context) {
+	return sendToMain('nunjuckRender', { templateName, context });
+}
+
+
+
+
+
 const templatesSource = window.templatesSource = new Map();
 templatesSource.set('backgroundTheme', '/templates/backgroundTheme');
 templatesSource.set('tabMover', '/templates/panel/tabMover');
-templatesSource.set('tabPageServerIp', '/templates/panel/tabPageServerIp');
 templatesSource.set('newTab', '/templates/newTab');
-templatesSource.set('tabUserStyles', '/templates/panel/tabUserStyles');